treewide: Move type converter utilities to separate subdirectory.
