+++ lib/vast/Conversion/CMakeLists.txt
@@ -2,4 +2,5 @@ add_subdirectory(ABI)
 add_subdirectory(PDLL)
 add_subdirectory(Common)
 add_subdirectory(FromHL)
+add_subdirectory(TypeConverters)
 add_subdirectory(Core)
