+++ lib/vast/Conversion/TypeConverters/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (c) 2022-present, Trail of Bits, Inc.
+
+add_vast_library(TypeConverters
+    TypeConverter.cpp
+)
