+++ lib/vast/Util/CMakeLists.txt
@@ -2,6 +2,5 @@
 
 add_vast_library(Util
     Region.cpp
-    TypeConverter.cpp
     Warnings.cpp
 )
